requirements.txt
fitz==0.0.1.dev2
nest_asyncio==1.6.0
python-dotenv==1.0.1
sentence_transformers==3.3.1
pdfplumber
openai
llama-index-core 
llama_index==0.12.8
llama-index-readers-file
llama-index-readers-docling 
llama-index-node-parser-docling  
llama-index-llms-bedrock
llama-index-llms-azure-openai
llama_index-embeddings-azure-openai
azure-storage-blob
azure-search-documents
fastapi
fastapi[all]
uvicorn
fastapi-cli
pydantic==2.10.6
# azure_storage==0.37.0